Reworked traps: much much cleaner now. There might be bugs, though.
Traps now remember how much ammo they have. The ammo quantities
(from David) are rather tentative.

Breaks savefiles.


git-svn-id: https://crawl-ref.svn.sourceforge.net/svnroot/crawl-ref/trunk@7076 c06c8d41-db1a-0410-9941-cceddc491573

+++ crawl-ref/source/hiscores.cc
@@ -1013,8 +1013,10 @@ void scorefile_entry::init()
 
     if (calc_item_values) // winners only
     {
-        points += (250000 * num_diff_runes)
-            * ((25000.0 * num_diff_runes) / (1+you.num_turns));
+        points +=
+            static_cast<long>(
+                (250000 * num_diff_runes)
+                * ((25000.0 * num_diff_runes) / (1+you.num_turns)));
     }
 
     // Players will have a hard time getting 1/10 of this (see XP cap):
